# src/main/resources/static/openApi.yml
openapi: 3.1.0
info:
  title: Movie Catalog API (IMDb Mini Clone)
  description: >
    A simple REST API to manage movies, including CRUD operations,
    filtering by genre, rating, title search, and sorting by rating.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local Development Server

tags:
  - name: Movies
    description: Operations related to movie resources.

paths:

  /api/movies:
    get:
      tags: [Movies]
      summary: Get all movies
      description: >
        Returns all movies. Supports filtering by genre, minimum rating,
        title search, and sorting by rating in descending order.
      parameters:
        - in: query
          name: genre
          schema:
            type: string
          description: Filter movies by genre.
        - in: query
          name: minRating
          schema:
            type: number
            format: float
          description: Filter movies with rating greater or equal to this value.
        - in: query
          name: search
          schema:
            type: string
          description: Search movies by partial title.
        - in: query
          name: sortByRating
          schema:
            type: boolean
          description: Sort movies by rating (high → low).
      responses:
        "200":
          description: List of movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movie"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Movies]
      summary: Create a new movie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieRequest"
      responses:
        "201":
          description: Movie created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/movies/{id}:
    get:
      tags: [Movies]
      summary: Get a movie by ID
      parameters:
        - $ref: "#/components/parameters/MovieId"
      responses:
        "200":
          description: Movie found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [Movies]
      summary: Update an existing movie
      parameters:
        - $ref: "#/components/parameters/MovieId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieRequest"
      responses:
        "200":
          description: Movie updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Movies]
      summary: Delete a movie
      parameters:
        - $ref: "#/components/parameters/MovieId"
      responses:
        "204":
          description: Movie deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

components:

  parameters:
    MovieId:
      name: id
      in: path
      required: true
      description: The ID of the movie.
      schema:
        type: integer
        format: int64

  schemas:

    Movie:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: The Dark Knight
        genre:
          type: string
          example: Action
        rating:
          type: number
          format: float
          example: 9.0
        releaseYear:
          type: integer
          example: 2008
        description:
          type: string
          example: Batman faces his most formidable foe, the Joker.
      required:
        - id
        - title

    MovieRequest:
      type: object
      properties:
        title:
          type: string
          example: Interstellar
        genre:
          type: string
          example: Sci-Fi
        rating:
          type: number
          format: float
          example: 8.6
        releaseYear:
          type: integer
          example: 2014
        description:
          type: string
          example: A team of explorers travel through a wormhole in space.
      required:
        - title

    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

  responses:

    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"